+++ Source/Core/VideoCommon/VideoCommon.vcxproj.filters
@@ -81,10 +81,19 @@
     <ClCompile Include="PixelShaderGen.cpp">
       <Filter>Shader Generators</Filter>
     </ClCompile>
+    <ClCompile Include="TexelDecoder.cpp">
+      <Filter>Base</Filter>
+    </ClCompile>
     <ClCompile Include="TextureConversionShader.cpp">
       <Filter>Shader Generators</Filter>
     </ClCompile>
-    <ClCompile Include="TextureDecoder_x64.cpp">
+    <ClCompile Include="TextureDecoderBase.cpp">
+      <Filter>Base</Filter>
+    </ClCompile>
+    <ClCompile Include="TextureDecoderHelpers.cpp">
+      <Filter>Base</Filter>
+    </ClCompile>
+    <ClCompile Include="TextureDecoderX86.cpp">
       <Filter>Shader Generators</Filter>
     </ClCompile>
     <ClCompile Include="VertexShaderGen.cpp">
@@ -178,7 +187,13 @@
     <ClInclude Include="OpcodeDecoding.h">
       <Filter>Decoding</Filter>
     </ClInclude>
+    <ClInclude Include="TexelDecoder.h">
+      <Filter>Decoding</Filter>
+    </ClInclude>
     <ClInclude Include="TextureDecoder.h">
+      <Filter>Decoding</Filter>
+    </ClInclude>
+    <ClInclude Include="TextureDecoderFunctions.h">
       <Filter>Decoding</Filter>
     </ClInclude>
     <ClInclude Include="BPFunctions.h">
